Common: Remove Exceptions
The satisfaction is immeasurable.
