+++ common/Windows/WinHostSys.cpp
@@ -19,7 +19,6 @@
 #include "common/RedtapeWindows.h"
 #include "common/Console.h"
 #include "common/General.h"
-#include "common/Exceptions.h"
 #include "common/StringUtil.h"
 #include "common/AlignedMalloc.h"
 #include "common/Assertions.h"
