+++ pcsx2-gsrunner/Main.cpp
@@ -28,7 +28,6 @@
 
 #include "common/Assertions.h"
 #include "common/Console.h"
-#include "common/Exceptions.h"
 #include "common/FileSystem.h"
 #include "common/MemorySettingsInterface.h"
 #include "common/Path.h"
