+++ pcsx2/CDVD/CDVDcommon.cpp
@@ -30,7 +30,6 @@
 #include "IsoFileFormats.h"
 
 #include "common/Assertions.h"
-#include "common/Exceptions.h"
 #include "common/FileSystem.h"
 #include "common/Path.h"
 #include "common/StringUtil.h"
