+++ pcsx2/CDVD/OutputIsoFile.cpp
@@ -18,7 +18,6 @@
 #include "CDVD/IsoFileFormats.h"
 #include "Host.h"
 
-#include "common/Exceptions.h"
 #include "common/FileSystem.h"
 #include "common/StringUtil.h"
